Responsive design for mobile phone
Fixed the overlapping of the nav buttons when viewing in mobile setting = input the header and it's content under the mobile query and adjusted the perimeters of how it will show from desktop to mobile view. However, due to the horizontal view on mobile, had to input the style for the mobile under the tablet view query (header, .header-container, nav a, .logo, .mybutton, and .container). Now it will work without issues when viewing from desktop to tablet/mobile (horizontal and vertical view). Therefore, need to ensure the coding is placed in proper area to work.