JQuery Example and applying to website
Learned the difference between Javascript and JQuery. JQuery is pretty much the abbreviation of the bigger programming language of Javascript. Input the JQuery to condense the nav menu into an icon so when clicked, it will show the nav bar menu text. Also, had to included modifying for margin, padding and colors.